% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mols.coefs}
\alias{mols.coefs}
\title{mols.coefs}
\usage{
mols.coefs(x, y, method = c("chol", "ginv", "svd", "qr"), tol = 1e-07)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Response matrix}

\item{method}{Coefficient calculation method. \code{chol} is the fastest
and \code{svd} is said to be the most reliable but maybe the slowest.}

\item{tol}{Tolerance setting for the \code{svd} method.}
}
\value{
Matrix of fit coefficients.
}
\description{
Multiple ordinary least squares coefficients. Used interally by
\link{mwlsr} to compute coefficients without weights.
}
