% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.coefStats}
\alias{mwlsr.coefStats}
\title{mwlsr.coefStats}
\usage{
mwlsr.coefStats(fit)
}
\arguments{
\item{fit}{mwlsr fit object}
}
\value{
mwlsr fit object with coefficient statistics results appended
}
\description{
Calculate coefficient standard errors, t-values and p-values. Results
are appended to the input mwlsr object.
}
