% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.contrastCoefficients}
\alias{mwlsr.contrastCoefficients}
\title{mwlsr.contrastCoefficients}
\usage{
mwlsr.contrastCoefficients(fit, contrast)
}
\arguments{
\item{fit}{mwlsr fit object}

\item{contrast}{Single-column contrast matrix}
}
\value{
mwlsr fit object with results appended
}
\description{
Calculate the coefficients of a contrast fit.
}
