% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.contrastModelMatrix}
\alias{mwlsr.contrastModelMatrix}
\title{mwlsr.contrastModelMatrix}
\usage{
mwlsr.contrastModelMatrix(design, contrast)
}
\arguments{
\item{design}{Full design matrix}

\item{contrast}{Single column contrast matrix indicating which levels
of the full design to contrast against one other.}
}
\value{
New, reduced, design matrix
}
\description{
Transforms a design matrix to a contrast matrix. Instead of calculating
the contrast result directly, as in \link{mwlsr.contrastTest}, 
this method would be used to create "reduced" model with a contrast
matrix and then you can evaluate significant associations with and LRT.
The code for this function is based on code within edgeR::glmLRT.
}
