% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.contrastTest}
\alias{mwlsr.contrastTest}
\title{mwlsr.contrastTest}
\usage{
mwlsr.contrastTest(fit, contrast = NULL, coef = NULL, ncomps = NULL,
  squeeze.var = FALSE)
}
\arguments{
\item{fit}{mwlsr fit object}

\item{contrast}{Single-column contrast matrix}

\item{coef}{Coefficient to test (for intercept models)}

\item{ncomps}{Sidak post-hoc correction factor. Default behavior is
no correction.}

\item{squeeze.var}{Employ limma's 'squeeze.var' method which not only
adjusts the model's dispersion but also the residual degrees of freedom}
}
\value{
data.frame with results of the test
}
\description{
Contrast one or more levels of the design factors against one or more
other levels. This kind of test uses the full model's dispersions
as a basis for the comparison of the means of two conditions. Useful for
comparing levels of a single multi-level factor, such as different 
groups in an RNA-Seq experiment, to one another to check for statistical
difference in means.
}
