% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.design2factor}
\alias{mwlsr.design2factor}
\title{mwlsr.design2factor}
\usage{
mwlsr.design2factor(X)
}
\arguments{
\item{X}{Design matrix}
}
\description{
Derives a single multi-level factor vector from a design matrix. 
This would produce senseless results for regression models.
}
