% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.Fstatistic}
\alias{mwlsr.Fstatistic}
\title{mwlsr.Fstatistic}
\usage{
mwlsr.Fstatistic(fit)
}
\arguments{
\item{fit}{mwlsr fit object}
}
\value{
Input mwlsr object with \code{F} and \code{F.pval} elements
attached
}
\description{
Calculate F-statistic for each model.
}
