% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.groupStats}
\alias{mwlsr.groupStats}
\title{mwlsr.groupStats}
\usage{
mwlsr.groupStats(fit)
}
\arguments{
\item{fit}{mwlsr fit object}
}
\value{
mwlsr fit object with group-level statistics appended
}
\description{
If your design was based on a single multi-level factor then you can
use this function to calculate per-group deviance, variance (dispersion), 
and standard error. Can be handy if you need to calculate group 
level variances.
}
