% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.LRT}
\alias{mwlsr.LRT}
\title{mwlsr.LRT}
\usage{
mwlsr.LRT(full.m, reduced.m, test = c("F", "LRT"))
}
\arguments{
\item{full.m}{Full model mwlsr object}

\item{reduced.m}{Reduced model mwlsr object}

\item{test}{Type of test to perform. For gaussian models, which is
all that mwlsr can do, you should use the F-test.}
}
\value{
data.frame with results of the test for all models.
}
\description{
Perform likelihood ratio test (LRT) between two models (typically a 
full model and a reduced model). Using an F based LRT on a full 
model compared to an intercept-only model should give the same
results as \link{mwlsr.Ftest}.
}
