% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.makeContrast}
\alias{mwlsr.makeContrast}
\title{mwlsr.makeContrast}
\usage{
mwlsr.makeContrast(y, lvls, x = NULL)
}
\arguments{
\item{y}{First level or levels to contrast.}

\item{lvls}{Either the factor levels or a full length factor vector or
the design matrix from the mwlsr object or the mwlsr object itself.}

\item{x}{Factor level or levels to contrast against those specified in 
\code{y}. If this is omitted then the level or levels specified in 
\code{y} are contrasted against all other levels.}
}
\value{
Single-column matrix specifying the contrast
}
\description{
Create a contrast matrix for computing statistical difference 
in means between levels of a design. For example if your model
is based on a single multi-level factor and you want to compare 
the average values of one level verses another within the context
of the full model then you'd specify those levels in this function
and then call on \link{mwlsr.contrastTest} to perform the test.
}
\details{
Levels specified for \code{y} or \code{x} must match levels
in the design matrix associated with the model you plan to perform 
the contrast test within.
}
