% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.overallFstatistic}
\alias{mwlsr.overallFstatistic}
\title{mwlsr.overallFstatistic}
\usage{
mwlsr.overallFstatistic(fit)
}
\arguments{
\item{fit}{mwlsr fit object}
}
\value{
Vector containing the results
}
\description{
Calculates F-statistic and p-value for all models. In this test we 
sum all of the residual deviance and compare it to the total sum of 
null deviance.
}
