% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.p.adjust}
\alias{mwlsr.p.adjust}
\title{mwlsr.p.adjust}
\usage{
mwlsr.p.adjust(p, n.comps = NA, n.levels = NA, n.samples = NA,
  method = c("sidak", "scheffe", "bonferroni"))
}
\arguments{
\item{p}{Vector of p-values}

\item{n.comps}{Number of contrasts being tested within the model. Defaults
to the total number of possible pairwise contrasts (probably excessive)}

\item{n.levels}{Number of levels in the factor design (or columns of the 
design matrix.}

\item{n.samples}{Required for "scheffe" method.}

\item{method}{Adjustment method. Sidak is the default.}
}
\value{
Adjusted p-values
}
\description{
Performs multi-contrast within model p-value adjustment. This adjustment
is performed at each p-value and is based on the number of contrasts 
being tested in the model.
}
