% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr}
\alias{mwlsr}
\title{mwlsr}
\usage{
mwlsr(data, design, weights = NULL, scale.weights = TRUE, data.err = NULL,
  coef.method = c("chol", "ginv", "svd", "qr"), coef.tol = 1e-07,
  coefs.only = FALSE)
}
\arguments{
\item{data}{Input response matrix with responses in columns}

\item{design}{Design matrix. See \link{model.matrix}}

\item{weights}{Weights matrix}

\item{scale.weights}{If TRUE then weights are scaled (default behavior)}

\item{data.err}{Additional per-response-value uncertainty that should be
considered in the final sum of squared residual. Useful if your response
values have some knowm measurement uncertainty that you'd like to 
have considered in the models.}

\item{coef.method}{Method used to compute coefficients. This setting is
passed to \link{mols.coefs} or \link{wls.coefs}}

\item{coef.tol}{Tolerance setting for svd based coefficient calculation.
Passed to \link{mols.coefs} or \link{wls.coefs}}

\item{coefs.only}{Stop at the coefficient calculation and return only 
the coefficients of the models.}
}
\value{
List with the following elements:
\item{coefficients}{Model coefficients}
\item{residuals}{Residuals of the fit}
\item{fitted.values}{Fitted values. Same dimension as the input response matrix.}
\item{deviance}{Sum of squared residuals}
\item{dispersion}{deviance / df.residual}
\item{null.deviance}{Sum of squared residuals for the NULL model (intercept only)}
\item{weights}{Weights matrix}
\item{prior.weights}{Weights matrix pre-scaling}
\item{weighted}{TRUE if fit was a weighted fit}
\item{df.residual}{Degrees of freedom of the model. \code{nrows(data) - ncol(design)}}
\item{df.null}{Degrees of freedom of the null model. \code{nrows(data) - 1}}
\item{y}{Input data matrix}
\item{y.err}{Input \code{data.err} matrix}
\item{X}{Design matrix}
\item{x}{If design matrix was based on factor levels then this will be a 
factor vector that matches the original grouping vector}
\item{intercept}{TRUE if the fit has an Intercept}
\item{coef.hat}{If the fit has an Intercept then this is a matrix of 
modified coefficients that represent the per-group averages. This is
calculated by adding the Intercept coefficients to each of the other 
coefficients. This only makes sense if your design was based on a single
multi-level factor}
}
\description{
Multiple Weighted Least Squares Regression (mwlsr). Used to fit gaussian
glm against multiple responses simultaneously.
}
\examples{
# Using the iris data.
design <- model.matrix(~Species, data=iris)
fit <- mwlsr(t(iris[, 1:4]), design)
# test data association with the Species factor
result <- mwlsr.Ftest(fit)
print(table(result$F.padj < 0.05))
}
