% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.rSquared}
\alias{mwlsr.rSquared}
\title{mwlsr.rSquared}
\usage{
mwlsr.rSquared(fit)
}
\arguments{
\item{fit}{Result of mwlsr fit}
}
\value{
Input mwlsr object (list) with \code{rquared} element attached
}
\description{
Calculate r-squared for each model in fit
}
