% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{mwlsr.tukeyHSD}
\alias{mwlsr.tukeyHSD}
\title{mwlsr.tukeyHSD}
\usage{
mwlsr.tukeyHSD(fit)
}
\arguments{
\item{fit}{mwlsr fit object}
}
\value{
list with everything in it (TODO: explain results)
}
\description{
Implementation of Tukey HSD per model. This only works for intercept 
designs.
}
