% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlsr.R
\name{wls.coefs}
\alias{wls.coefs}
\title{wls.coefs}
\usage{
wls.coefs(x, y, weights = NULL, method = c("chol", "ginv", "svd", "qr"),
  tol = 1e-07)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Response vector}

\item{weights}{Weights vector.}

\item{method}{Coefficient calculation method. See \link{mols.coefs}.}

\item{tol}{Tolerance for \code{svd} coefficient method.}
}
\value{
Vector of fit coefficients
}
\description{
Calculates coefficients for a single response with weights.
}
